requirements.txt
torch
torchaudio
librosa
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
sentence-transformers
umap-learn